refactor [upload-publication.js]: move upload lohics in modul
